[nacho] only set write dominated when size = line size
We can only set the write dominated bit (that later prevents setting
the possible war bit) if ALL the bytes in the line are written.

Otherwise a scenario can occur where 2 bytes are written 0,1. This
causes the write dominated bit to be set for the line. Now, later, the
two other bytes in the line are Read (2,3).
During the read, the read dominated flag is not set, and neither is the
possible WAR flag, as it thinks the line is write dominated. However, if
we now write to bytes 2,3 in the cache, and later evict bytes 2,3 to NVM,
the system will think they won't cause a WAR, so they are written back freely.
But this is a bug, as there is a WAR to byes 2,3.

So the solution is to ONLY set the write dominated flag is the WHOLE
line is written to.